RANCHER-2414. Refactor authentication header methods to utilize the User object for tenant operations and align them with the current approach.
