+++ src/org/folio/rest_v2/eureka/Keycloak.groovy
@@ -3,6 +3,7 @@ package org.folio.rest_v2.eureka
 import groovy.text.StreamingTemplateEngine
 import hudson.util.Secret
 import org.folio.models.EurekaTenant
+import org.folio.models.User
 
 /**
  * The Keycloak class is responsible for various operations related to
@@ -59,12 +60,20 @@ class Keycloak extends Base {
     return getAuthorizedHeaders(getAuthTenantToken(tenant), addOkapiAuth) + ["X-Okapi-Tenant": tenant.tenantId]
   }
 
+  Map<String,String> getAuthTenantHeaders(EurekaTenant tenant, User user, boolean addOkapiAuth = false) {
+    return getAuthorizedHeaders(getAuthTenantToken(tenant, user), addOkapiAuth) + ["X-Okapi-Tenant": tenant.tenantId]
+  }
+
   String getAuthMasterTenantToken() {
     return getAuthToken("master", MASTER_TENANT_CLIENT_ID, MASTER_TENANT_CLIENT_SECRET)
   }
 
   String getAuthTenantToken(EurekaTenant tenant) {
     return getAuthToken(tenant.tenantId, tenant.clientId, tenant.clientSecret)
+  }
+
+  String getAuthTenantToken(EurekaTenant tenant, User user) {
+    return getAuthUserToken(tenant.tenantId, user.username, user.password)
   }
 
   String getAuthToken(String tenantId, String clientId, Secret clientSecret){
@@ -80,6 +89,22 @@ class Keycloak extends Base {
     logger.info("Access token obtained successfully from Keycloak service")
 
     return response['access_token']
+  }
+
+  String getAuthUserToken(String tenantId, String username, Secret password) {
+    logger.info("Getting access token from Keycloak service for user " + username)
+
+    String url = generateUrl("/authn/login")
+
+    Map<String, String> headers = ['Content-Type': 'application/json', 'x-okapi-tenant': tenantId]
+
+    Map requestBody = ["username": username, "password": password.getPlainText()]
+
+    def response = restClient.post(url, requestBody, headers).body
+
+    logger.info("Access token obtained successfully from Keycloak service for user " + username)
+
+    return response['okapiToken']
   }
 
   Keycloak defineTTL(String tenantId, int ttl = 300) {
