+++ src/org/folio/rest_v2/eureka/Kong.groovy
@@ -1,7 +1,7 @@
 package org.folio.rest_v2.eureka
 
-import hudson.util.Secret
 import org.folio.models.EurekaTenant
+import org.folio.models.User
 
 class Kong extends Base {
 
@@ -48,23 +48,7 @@ class Kong extends Base {
     return getDefaultHeaders() + keycloak.getAuthTenantHeaders(tenant, addOkapiAuth)
   }
 
-  Map<String, String> getTenantHttpHeaders(String tenantId, String token, boolean addOkapiAuth = false) {
-    return getDefaultHeaders() + keycloak.getAuthorizedHeaders(token, addOkapiAuth) + ["X-Okapi-Tenant": tenantId]
-  }
-
-  String getAuthUserToken(String tenantId, String username, Secret password) {
-    logger.info("Getting access token via login for user " + username)
-
-    String url = generateUrl("/authn/login")
-
-    Map<String, String> headers = ['Content-Type': 'application/json', 'x-okapi-tenant': tenantId]
-
-    Map requestBody = ["username": username, "password": password.getPlainText()]
-
-    def response = restClient.post(url, requestBody, headers).body
-
-    logger.info("Access token obtained successfully via login for user " + username)
-
-    return response['okapiToken']
+  Map<String, String> getTenantUserHttpHeaders(EurekaTenant tenant, User user = tenant.getAdminUser(), boolean addOkapiAuth = false) {
+    return getDefaultHeaders() + keycloak.getAuthTenantHeaders(tenant, user, addOkapiAuth)
   }
 }
